requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Binance API Integration (Primary Trading API)
python-binance>=1.0.19

# Web Dashboard (Flask-based interactive web interface)
Flask>=3.0.0

# Dashboard and Visualization (for programmatic charts and exports)
matplotlib>=3.7.0
plotly>=5.18.0

# Optional: For View Session dashboard (Streamlit)
# streamlit>=1.28.0  # For interactive dashboard
# Install with: pip install streamlit

# Optional: For extended features
# seaborn>=0.12.0    # For visualizations
# scipy>=1.10.0      # For statistical analysis